--- a/app/client/craco.dev.config.js
+++ b/app/client/craco.dev.config.js
@@ -7,11 +7,18 @@ module.exports = merge(common, {
     client: {
       overlay: {
         warnings: false,
-        errors: false
-      }
-    }
+        errors: false,
+      },
+    },
   },
   optimization: {
     minimize: false,
+  },
+  cache: {
+    type: "filesystem",
+    memoryCacheUnaffected: true,
+  },
+  experiments: {
+    cacheUnaffected: true,
   },
 });