## Description
Dev mode performance:
- Disabled background tape checks for dev builds briefly.
- Configured caching for dev mode
- Added TS exclude paths

Timings: 
before:  build ~ 120 sec, recompile ~ 115 sec
after: build ~ 16 sec, recompile ~ 8 sec